update shopping cart & login

+++ src/main/java/asm/dao/OrderDAO.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import asm.entity.Orders;
 
 
-public interface OrderDAO extends JpaRepository<Orders, Integer>{
+
+public interface OrderDAO extends JpaRepository<Orders, Long> {
 
 }
